requirements.txt
streamlit==1.35.0
speechbrain==0.5.16
torch==2.2.2
torchaudio==2.2.2
pydub==0.25.1
scikit-learn==1.4.2
langdetect==1.0.9
moviepy
matplotlib
torchvision
soundfile
torchaudio
openai-whisper